# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HyperLocal India Super App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── Feedback ─────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    UserProfile:
      type: object
      required: [id, replitId, role]
      properties:
        id:
          type: integer
        replitId:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, seller, admin]
        profileImageUrl:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    UpdateUserProfileBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
...